# contracts/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for the Reverse Marketplace platform",
    "title": "Reverse Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/reviews/sellers/{sellerId}": {
      "get": {
        "description": "List seller reviews",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Results per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Sort order",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "newest",
              "enum": [
                "newest",
                "oldest",
                "highest",
                "lowest"
              ],
              "type": "string"
            }
          },
          {
            "description": "Seller profile UUID",
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
...
    "/api/v1/saved-sellers/": {
      "get": {
        "description": "List your saved sellers",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Results per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Saved Sellers"
        ]
      }
    },
    "/api/v1/saved-sellers/{sellerId}": {
      "delete": {
        "description": "Remove a seller from favorites",
        "parameters": [
          {
            "description": "Seller profile UUID",
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Saved Sellers"
        ]
      },
      "post": {
        "description": "Save a seller to your favorites",
        "parameters": [
          {
            "description": "Seller profile UUID",
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Saved Sellers"
        ]
      }
    },
...
    "/api/v1/sellers/": {
      "post": {
        "description": "Create seller profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "bio": {
                    "description": "Business bio (e.g. Licensed plumber serving DFW for 10 years)",
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "businessHours": {
                    "additionalProperties": {
                      "properties": {
                        "close": {
                          "description": "Closing time HH:MM (e.g. 17:00)",
                          "pattern": "^\\d{2}:\\d{2}$",
                          "type": "string"
                        },
                        "closed": {
                          "description": "true if closed this day",
                          "type": "boolean"
                        },
                        "open": {
                          "description": "Opening time HH:MM (e.g. 08:00)",
                          "pattern": "^\\d{2}:\\d{2}$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "open",
                        "close"
                      ],
                      "type": "object"
                    },
                    "description": "Business hours by day of week",
                    "nullable": true,
                    "type": "object"
                  },
                  "businessName": {
                    "description": "Business name (e.g. Dallas Plumbing Services)",
                    "maxLength": 255,
                    "type": "string"
                  },
                  "businessWebsite": {
                    "description": "Website URL (e.g. https://myplumbing.com)",
                    "format": "uri",
                    "maxLength": 500,
                    "nullable": true,
                    "type": "string"
                  },
                  "categories": {
                    "default": [],
                    "description": "Array of category UUIDs you serve",
                    "items": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "serviceRadiusMiles": {
                    "default": 25,
                    "description": "Service radius in miles (default 25)",
                    "maximum": 500,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "subcategories": {
                    "default": [],
                    "description": "Array of subcategory UUIDs",
                    "items": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "yearsExperience": {
                    "description": "Years of experience (e.g. 10)",
                    "maximum": 100,
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Sellers"
        ]
      }
    },
...
    "/api/v1/sellers/{sellerId}": {
      "get": {
        "description": "Get public seller profile",
        "parameters": [
          {
            "description": "Seller profile UUID",
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "Sellers"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Reverse Marketplace API (base path)",
      "url": "/api/v1"
    }
  ]
}